+++ shimmer/__init__.py
@@ -2,10 +2,7 @@ from shimmer.dataset import RepeatedDataset
 from shimmer.modules.contrastive_loss import (
     ContrastiveLoss,
     ContrastiveLossType,
-    ContrastiveLossWithUncertainty,
-    ContrastiveLossWithUncertaintyType,
     contrastive_loss,
-    contrastive_loss_with_uncertainty,
 )
 from shimmer.modules.domain import DomainModule, LossOutput
 from shimmer.modules.global_workspace import (
@@ -20,7 +17,6 @@ from shimmer.modules.gw_module import (
     GWDecoder,
     GWEncoder,
     GWEncoderLinear,
-    GWEncoderWithUncertainty,
     GWModule,
     GWModuleBase,
     GWModuleWithUncertainty,
@@ -31,17 +27,17 @@ from shimmer.modules.losses import (
     GWLossesWithUncertainty,
     LossCoefs,
 )
+from shimmer.modules.selection import (
+    RandomSelection,
+    SelectionBase,
+    SingleDomainSelection,
+)
 from shimmer.modules.utils import (
     batch_cycles,
-    batch_cycles_with_uncertainty,
     batch_demi_cycles,
-    batch_demi_cycles_with_uncertainty,
     batch_translations,
-    batch_translations_with_uncertainty,
     cycle,
-    cycle_with_uncertainty,
     translation,
-    translation_with_uncertainty,
 )
 from shimmer.types import (
     LatentsDomainGroupDT,
@@ -79,16 +75,12 @@ __all__ = [
     "GWDecoder",
     "GWEncoder",
     "GWEncoderLinear",
-    "GWEncoderWithUncertainty",
     "GWModuleBase",
     "GWModule",
     "GWModuleWithUncertainty",
     "ContrastiveLossType",
-    "ContrastiveLossWithUncertaintyType",
     "contrastive_loss",
     "ContrastiveLoss",
-    "contrastive_loss_with_uncertainty",
-    "ContrastiveLossWithUncertainty",
     "LossCoefs",
     "GWLossesBase",
     "GWLosses",
@@ -99,12 +91,10 @@ __all__ = [
     "batch_translations",
     "cycle",
     "translation",
-    "cycle_with_uncertainty",
-    "translation_with_uncertainty",
-    "batch_translations_with_uncertainty",
-    "batch_demi_cycles_with_uncertainty",
-    "batch_cycles_with_uncertainty",
     "MIGRATION_DIR",
     "migrate_model",
     "SaveMigrations",
+    "RandomSelection",
+    "SelectionBase",
+    "SingleDomainSelection",
 ]