+++ shimmer/modules/__init__.py
@@ -20,7 +20,6 @@ from shimmer.modules.gw_module import (
     GWDecoder,
     GWEncoder,
     GWEncoderLinear,
-    GWEncoderWithUncertainty,
     GWModule,
     GWModuleBase,
     GWModuleWithUncertainty,
@@ -31,17 +30,17 @@ from shimmer.modules.losses import (
     GWLossesWithUncertainty,
     LossCoefs,
 )
+from shimmer.modules.selection import (
+    RandomSelection,
+    SelectionBase,
+    SingleDomainSelection,
+)
 from shimmer.modules.utils import (
     batch_cycles,
-    batch_cycles_with_uncertainty,
     batch_demi_cycles,
-    batch_demi_cycles_with_uncertainty,
     batch_translations,
-    batch_translations_with_uncertainty,
     cycle,
-    cycle_with_uncertainty,
     translation,
-    translation_with_uncertainty,
 )
 from shimmer.modules.vae import (
     VAE,
@@ -64,7 +63,6 @@ __all__ = [
     "GWDecoder",
     "GWEncoder",
     "GWEncoderLinear",
-    "GWEncoderWithUncertainty",
     "GWModuleBase",
     "GWModule",
     "GWModuleWithUncertainty",
@@ -90,9 +88,7 @@ __all__ = [
     "batch_translations",
     "cycle",
     "translation",
-    "cycle_with_uncertainty",
-    "translation_with_uncertainty",
-    "batch_translations_with_uncertainty",
-    "batch_demi_cycles_with_uncertainty",
-    "batch_cycles_with_uncertainty",
+    "RandomSelection",
+    "SelectionBase",
+    "SingleDomainSelection",
 ]